Remove problematic GOROOT export.
This causes the istioctl upload to fail when it can't find the GOROOT
properly.  Since we're moving this to build in a container there's no
reason for suck hackery.  We'll rely on the GOROOT in the env, which is
not prone to breakage.

find: '/tmp/gopath/src/istio.io/pilot/bazel-pilot/external': No such
file or directory